Attaching error handler to blueprint level instead of global. Issue #362
